+++ R/rollr2-package.R
@@ -2,6 +2,14 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom ggplot2 aes
+#' @importFrom ggplot2 geom_col
+#' @importFrom ggplot2 ggplot
+#' @importFrom ggplot2 labs
+#' @importFrom ggplot2 scale_fill_manual
+#' @importFrom ggplot2 scale_x_continuous
+#' @importFrom ggplot2 theme_minimal
+#' @importFrom rlang .data
 #' @importFrom rlang abort
 ## usethis namespace: end
 NULL